requirements.txt
Flask==2.3.3
flask-cors==4.0.0
PyGithub==2.1.1
requests==2.31.0